# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "智慧餐厅saas服务.",
        "title": "MP-ISV API",
        "contact": {
            "name": "codeCloud",
            "url": "https://codecloud.ltd",
            "email": "codeCloud2020@163.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:4002",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/settings/recharge": {
            "get": {
                "description": "查看充值规则设置",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "restaurant 充值规则"
                ],
                "summary": "充值规则设置",
                "responses": {
                    "200": {
                        "description": "code:1 =\u003e 获取成功，code:0 =\u003e 获取失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gincmf_plugins_restaurantPlugin_model.ReturnData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Recharge"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"登录状态已失效！\"}",
                        "schema": {
                            "$ref": "#/definitions/gincmf_plugins_restaurantPlugin_model.ReturnData"
                        }
                    }
                }
            },
            "post": {
                "description": "查看充值规则设置",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "restaurant 充值规则"
                ],
                "summary": "充值规则设置",
                "responses": {
                    "200": {
                        "description": "code:1 =\u003e 获取成功，code:0 =\u003e 获取失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gincmf_plugins_restaurantPlugin_model.ReturnData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Recharge"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"登录状态已失效！\"}",
                        "schema": {
                            "$ref": "#/definitions/gincmf_plugins_restaurantPlugin_model.ReturnData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gincmf_plugins_restaurantPlugin_model.ReturnData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.Recharge": {
            "type": "object",
            "properties": {
                "gear": {
                    "description": "储值金额档位",
                    "type": "number"
                },
                "money": {
                    "type": "number"
                },
                "money_enabled": {
                    "type": "integer"
                },
                "point": {
                    "type": "integer"
                },
                "point_enabled": {
                    "type": "integer"
                },
                "voucher": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.voucherItem"
                    }
                },
                "voucher_enabled": {
                    "type": "integer"
                }
            }
        },
...
        "model.voucherItem": {
            "type": "object",
            "properties": {
                "send_type": {
                    "description": "发放方式：once：发放1次 month：每月发放",
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                },
                "voucher_id": {
                    "type": "integer"
                },
                "voucher_name": {
                    "type": "string"
                }
            }
        },
...
    }
}